pretty it up and added service classes
