+++ app/src/main/res/layout/activity_calendar.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".Activity.MainActivity">
+    tools:context=".Activity.CalendarActivity">
 
     <!--<android.support.design.widget.AppBarLayout-->
     <!--android:layout_width="match_parent"-->
@@ -13,6 +13,14 @@
     <!--android:theme="@style/AppTheme.AppBarOverLay">-->
     <!--</android.support.design.widget.AppBarLayout>-->
 
-    <include layout="@layout/content_main" />
+    <include layout="@layout/content_calendar" />
+
+    <android.support.design.widget.FloatingActionButton
+        android:id="@+id/calendar_sync_fab"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_gravity="bottom|end"
+        android:layout_margin="16dp"
+        app:srcCompat="@android:drawable/ic_popup_sync"/>
 
 </android.support.design.widget.CoordinatorLayout>
