--- app/src/main/res/layout/date_content.xml
@@ -1,65 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    android:orientation="vertical"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent">
-    <TextView
-        android:id="@+id/date_view_date_tv"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_alignParentEnd="true"
-        android:textStyle="bold"
-        android:textSize="20sp"
-        android:text="@string/date_content_date_tv_holder"/>
-    <LinearLayout
-        android:id="@+id/date_content_content_layout"
-        android:layout_below="@id/date_view_date_tv"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:weightSum="1"
-        android:orientation="vertical">
-        <RelativeLayout
-            android:id="@+id/date_content_facebook_content_layout"
-            android:layout_width="match_parent"
-            android:layout_height="0dp"
-            android:layout_weight="1">
-            <ImageView
-                android:id="@+id/date_content_facebook_events_iv"
-                android:layout_width="wrap_content"
-                android:layout_height="30dp"
-                android:background="@color/com_facebook_blue"
-                android:src="@drawable/date_content_facebook_event_banner"/>
-            <android.support.v7.widget.RecyclerView
-            android:id="@+id/date_content_event_list"
-            android:layout_below="@+id/date_content_facebook_events_iv"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:scrollbars="vertical" />
-        </RelativeLayout>
-        <RelativeLayout
-            android:id="@+id/date_content_twitter_content_layout"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content">
-            <ImageView
-                android:id="@+id/date_content_twitter_tweets_iv"
-                android:layout_width="wrap_content"
-                android:layout_height="30dp"
-                android:background="@color/colorTwitterLightBlue"
-                android:src="@drawable/date_content_twitter_tweet_banner"/>
-    <android.support.v7.widget.RecyclerView
-        android:id="@+id/date_content_twitter_tweet_list"
-        android:layout_below="@id/date_content_twitter_tweets_iv"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:scrollbars="vertical" />
-            <ImageView
-                android:id="@+id/date_content_google_events_iv"
-                android:layout_width="wrap_content"
-                android:layout_height="30dp"
-                android:layout_below="@+id/date_content_twitter_tweet_list"
-                android:background="@color/colorGoogleRed"
-                android:src="@drawable/date_content_google_event_banner"/>
-        </RelativeLayout>
-    </LinearLayout>
-</RelativeLayout>
