+++ app/src/main/res/layout/item_fb_event.xml
@@ -1,7 +1,8 @@
 <?xml version="1.0" encoding="utf-8"?>
 <RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
     android:layout_width="match_parent"
-    android:layout_height="wrap_content">
+    android:layout_height="wrap_content"
+    android:padding="10dp">
     <RelativeLayout
         android:id="@+id/fb_event_item_iv_layout"
         android:layout_width="wrap_content"
