+++ app/src/main/res/values/colors.xml
@@ -5,4 +5,6 @@
     <color name="colorAccent">#FF4081</color>
     <color name="colorTwitterLightBlue">#33CCFF</color>
     <color name="colorGoogleRed">#D85334</color>
+    <color name="colorBlack">#000000</color>
+    <color name="colorWhite">#FFFFFF</color>
 </resources>
